+++ CHANGELOG.md
@@ -11,3 +11,10 @@
 ## [1.2.1] - 2024-09-01
 
 ## [1.2.2] - 2024-10-05
+
+## [2.0.0] - 2024-10-06
+
+INTERFACE CHANGED:
+
+- client.getQueue now takes device name as argument instead of serial/type.
+- client.getDevices no longer filters devices.